requirements.txt
django==1.5.2
MySQL-python==1.2.5
djangorestframework==2.4.4
simplejson==3.3.1
python-dateutil==1.5
requests==2.4.1
django-crispy-forms==1.4.0
xlwt==0.7.5
django-xadmin==0.4.2
pil==1.1.7
qrcode==5.1
pybcs==1.2.91